+++ Tintoreria/Core/Trabajo.cs
@@ -10,7 +10,7 @@ namespace Upds.Sistemas.ProgWeb2.Tintoreria.Core
         public int IdTrabajo { get; set; }
         public Cliente Cliente { get; set; }
         public DateTime FechaTrabajo { get; set; }
-        public double TotalPrecio { get; set; }
+        public decimal TotalPrecio { get; set; }
         public DateTime FechaEntrega { get; set; }
         public Pedido PedidoDistancia { get; set; }
         public bool EntregaDomicilio { get; set; }
